feat : 공고글작성 1단계 낮춤
기존 : template단에서 data fetching 및 post요청
이 후 :  organism단에서 수행
