+++ src/pages/DetailNotificationPage.tsx
@@ -1,19 +1,7 @@
-import DescriptionBoxNoti from '../components/atoms/DescriptionBoxNoti';
-import DescriptionBoxTitle from '../components/atoms/DescriptionBoxTitle';
-import BottomNavBar from '../components/molecules/BottomNavBar';
 import DetailNotificationTemplate from '../components/templates/DetailNotificationTemplate';
-import * as S from '../styles/pages/DescriptiontBoxContainer';
 
-const DetailNotification = () => {
-  return (
-    <S.Container>
-      <DescriptionBoxTitle title="공고글 제목" />
-      <DescriptionBoxNoti>
-        <DetailNotificationTemplate />
-      </DescriptionBoxNoti>
-      <BottomNavBar />
-    </S.Container>
-  );
+const DetailNotificationPage = () => {
+  return <DetailNotificationTemplate />;
 };
 
-export default DetailNotification;
+export default DetailNotificationPage;
