+++ src/styles/organisms/DogProfile.ts
@@ -4,6 +4,7 @@ export const Container = styled.div`
   flex-direction: column;
   width: 100%;
   /* height: 45%; */
+  margin-top: 1rem;
   padding-bottom: 0.5rem;
   border-bottom: 1px solid #e2e2e2;
 `;
